SW-4933 Merge the redundant Page components
- refactored deliverables view to work with the OG Page
