+++ src/components/BreadCrumbs/index.tsx
@@ -1,10 +1,7 @@
-import React, { useRef } from 'react';
-import { Box, Grid, Typography, useTheme } from '@mui/material';
+import React from 'react';
+import { Box, Typography, useTheme } from '@mui/material';
 import BackToLink from 'src/components/common/BackToLink';
 import Link from 'src/components/common/Link';
-import TfMain from 'src/components/common/TfMain';
-import PageSnackbar from 'src/components/PageSnackbar';
-import PageHeaderWrapper from 'src/components/common/PageHeaderWrapper';
 
 export type Crumb = {
   name: string; // name of crumb
@@ -49,55 +46,5 @@ export default function BreadCrumbs({ hierarchical, crumbs }: BreadCrumbsProps):
         </Box>
       ))}
     </Box>
-  );
-}
-
-export type PageProps = {
-  crumbs: Crumb[];
-  title: string;
-  children: React.ReactNode;
-  rightComponent?: React.ReactNode;
-};
-
-/**
- * A generic page structure with bread crumbs, title, header wrapper and instantiated children.
- */
-export function Page({ crumbs, title, children, rightComponent }: PageProps): JSX.Element {
-  const contentRef = useRef(null);
-  const theme = useTheme();
-
-  return (
-    <TfMain>
-      <PageHeaderWrapper nextElement={contentRef.current}>
-        <BreadCrumbs crumbs={crumbs} hierarchical={true} />
-        <Grid container justifyContent={'space-between'}>
-          <Grid item xs={8}>
-            <Typography
-              sx={{
-                marginTop: theme.spacing(3),
-                marginBottom: theme.spacing(4),
-                paddingLeft: theme.spacing(3),
-                fontSize: '20px',
-                fontWeight: 600,
-                color: theme.palette.TwClrBaseGray800,
-              }}
-            >
-              {title}
-            </Typography>
-          </Grid>
-          {rightComponent && (
-            <Grid item xs={4} sx={{ textAlign: 'right' }}>
-              {rightComponent}
-            </Grid>
-          )}
-          <Grid item xs={12}>
-            <PageSnackbar />
-          </Grid>
-        </Grid>
-      </PageHeaderWrapper>
-      <Grid container ref={contentRef}>
-        {children}
-      </Grid>
-    </TfMain>
   );
 }
