+++ src/components/ProjectView/index.tsx
@@ -3,7 +3,8 @@ import { useParams } from 'react-router-dom';
 import { useHistory } from 'react-router';
 import { Grid } from '@mui/material';
 import { DropdownItem } from '@terraware/web-components';
-import { Crumb, Page } from 'src/components/BreadCrumbs';
+import { Crumb } from 'src/components/BreadCrumbs';
+import Page from 'src/components/Page';
 import Card from 'src/components/common/Card';
 import strings from 'src/strings';
 import { APP_PATHS } from 'src/constants';
