+++ src/scenes/AcceleratorRouter/AcceleratorDeliverableView.tsx
@@ -1,11 +1,15 @@
-import React from 'react';
-import { BusySpinner, Message } from '@terraware/web-components';
+import React, { useMemo } from 'react';
+import { BusySpinner } from '@terraware/web-components';
 import strings from 'src/strings';
+import { APP_PATHS } from 'src/constants';
 import useDeviceInfo from 'src/utils/useDeviceInfo';
+import { useLocalization } from 'src/providers';
 import Card from 'src/components/common/Card';
+import { Crumb } from 'src/components/BreadCrumbs';
 import Page from 'src/components/Page';
 import AcceleratorDocumentsList from 'src/scenes/AcceleratorRouter/AcceleratorDocumentsList';
 import { EditProps, ViewProps } from 'src/components/DeliverableView/types';
+import MobileMessage from 'src/components/DeliverableView/MobileMessage';
 import TitleBar from 'src/components/DeliverableView/TitleBar';
 import DocumentsUploader from 'src/components/DeliverableView/DocumentsUploader';
 import Metadata from 'src/components/DeliverableView/Metadata';
@@ -17,30 +21,33 @@ export type Props = EditProps & {
 const AcceleratorDeliverableView = (props: Props): JSX.Element => {
   const { ...viewProps }: ViewProps = props;
   const { isMobile } = useDeviceInfo();
+  const { activeLocale } = useLocalization();
+
+  const crumbs: Crumb[] = useMemo(
+    () => [
+      {
+        name: activeLocale ? strings.DELIVERABLES : '',
+        to: APP_PATHS.ACCELERATOR_DELIVERABLES,
+      },
+    ],
+    [activeLocale]
+  );
 
   if (isMobile) {
-    return (
-      <Page
-        content={<Message body={strings.FEATURE_AVAILABLE_ON_DESKTOP} priority='info' type='page' />}
-        title={<TitleBar deliverable={props.deliverable} isAcceleratorConsole={props.isAcceleratorConsole} />}
-      />
-    );
+    return <MobileMessage {...viewProps} />;
   }
 
   return (
-    <Page
-      content={
-        <>
-          {props.isBusy && <BusySpinner />}
-          <Card style={{ display: 'flex', flexDirection: 'column' }}>
-            <Metadata {...viewProps} />
-            <DocumentsUploader {...viewProps} />
-            <AcceleratorDocumentsList {...viewProps} />
-          </Card>
-        </>
-      }
-      title={<TitleBar {...props} />}
-    />
+    <Page title={<TitleBar {...viewProps} />} rightComponent={props.callToAction} crumbs={crumbs}>
+      <>
+        {props.isBusy && <BusySpinner />}
+        <Card style={{ display: 'flex', flexDirection: 'column', flexGrow: 1 }}>
+          <Metadata {...viewProps} />
+          <DocumentsUploader {...viewProps} />
+          <AcceleratorDocumentsList {...viewProps} />
+        </Card>
+      </>
+    </Page>
   );
 };
