+++ src/scenes/DeliverablesRouter/DeliverableView.tsx
@@ -1,14 +1,18 @@
-import React from 'react';
-import { BusySpinner, Message } from '@terraware/web-components';
+import React, { useMemo } from 'react';
+import { BusySpinner } from '@terraware/web-components';
 import strings from 'src/strings';
+import { APP_PATHS } from 'src/constants';
 import useDeviceInfo from 'src/utils/useDeviceInfo';
+import { useLocalization } from 'src/providers';
 import Card from 'src/components/common/Card';
+import { Crumb } from 'src/components/BreadCrumbs';
 import Page from 'src/components/Page';
 import DocumentsList from 'src/scenes/DeliverablesRouter/DocumentsList';
 import { EditProps, ViewProps } from 'src/components/DeliverableView/types';
 import TitleBar from 'src/components/DeliverableView/TitleBar';
 import DocumentsUploader from 'src/components/DeliverableView/DocumentsUploader';
 import Metadata from 'src/components/DeliverableView/Metadata';
+import MobileMessage from 'src/components/DeliverableView/MobileMessage';
 
 export type Props = EditProps & {
   isBusy?: boolean;
@@ -17,30 +21,33 @@ export type Props = EditProps & {
 const DeliverableView = (props: Props): JSX.Element => {
   const { ...viewProps }: ViewProps = props;
   const { isMobile } = useDeviceInfo();
+  const { activeLocale } = useLocalization();
+
+  const crumbs: Crumb[] = useMemo(
+    () => [
+      {
+        name: activeLocale ? strings.DELIVERABLES : '',
+        to: APP_PATHS.DELIVERABLES,
+      },
+    ],
+    [activeLocale]
+  );
 
   if (isMobile) {
-    return (
-      <Page
-        content={<Message body={strings.FEATURE_AVAILABLE_ON_DESKTOP} priority='info' type='page' />}
-        title={<TitleBar deliverable={props.deliverable} isAcceleratorConsole={props.isAcceleratorConsole} />}
-      />
-    );
+    return <MobileMessage {...viewProps} />;
   }
 
   return (
-    <Page
-      content={
-        <>
-          {props.isBusy && <BusySpinner />}
-          <Card style={{ display: 'flex', flexDirection: 'column' }}>
-            <Metadata {...viewProps} />
-            <DocumentsUploader {...viewProps} />
-            <DocumentsList {...viewProps} />
-          </Card>
-        </>
-      }
-      title={<TitleBar {...props} />}
-    />
+    <Page title={<TitleBar {...props} />} crumbs={crumbs}>
+      <>
+        {props.isBusy && <BusySpinner />}
+        <Card style={{ display: 'flex', flexDirection: 'column', flexGrow: 1 }}>
+          <Metadata {...viewProps} />
+          <DocumentsUploader {...viewProps} />
+          <DocumentsList {...viewProps} />
+        </Card>
+      </>
+    </Page>
   );
 };
