+++ src/stories/BreadCrumbs.stories.tsx
@@ -1,14 +1,8 @@
 import { Story } from '@storybook/react';
-import BreadCrumbs, { BreadCrumbsProps, Page, PageProps } from 'src/components/BreadCrumbs';
-import Button from 'src/components/common/button/Button';
-import OptionsMenu from 'src/components/common/OptionsMenu';
+import BreadCrumbs, { BreadCrumbsProps } from 'src/components/BreadCrumbs';
 
 const BreadCrumbsTemplate: Story<BreadCrumbsProps> = (args: BreadCrumbsProps) => {
   return <BreadCrumbs {...args} />;
-};
-
-const PageTemplate: Story<PageProps> = (args: PageProps) => {
-  return <Page {...args} />;
 };
 
 export default {
@@ -59,59 +53,4 @@ NonHierarchicalCrumbs.args = {
       to: '/xyz/house',
     },
   ],
-};
-
-export const PageWithCrumbs = PageTemplate.bind({});
-
-PageWithCrumbs.args = {
-  crumbs: [
-    {
-      name: 'root',
-      to: '/root',
-    },
-    {
-      name: 'child',
-      to: '/child',
-    },
-  ],
-  title: 'hello world',
-  children: 'Test page',
-};
-
-export const PageWithCrumbsAndRightComponent = PageTemplate.bind({});
-
-PageWithCrumbsAndRightComponent.args = {
-  crumbs: [
-    {
-      name: 'root',
-      to: '/root',
-    },
-    {
-      name: 'child',
-      to: '/child',
-    },
-  ],
-  title: 'hello world',
-  children: 'Test page',
-  rightComponent: (
-    <>
-      <Button
-        label={'Click this button'}
-        icon='iconEdit'
-        onClick={() => {
-          // tslint:disable-next-line:no-console
-          console.log('Thank you for clicking me');
-        }}
-        size='medium'
-        id='button'
-      />
-      <OptionsMenu
-        onOptionItemClick={(item) => {
-          // tslint:disable-next-line:no-console
-          console.log('You clicked on ', item);
-        }}
-        optionItems={[{ label: 'Click this option', value: 'click' }]}
-      />
-    </>
-  ),
 };
